// frontend/src/api/generated/models/AdminDepartmentResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AdminDepartmentResponse = {
    created_at: string;
    id: string;
    name: string;
    parent_id: (string | null);
    path: string;
    sort_order: number;
    status: 'active' | 'disabled';
    tenant_id: string;
    updated_at: string;
    version: number;
};